Progress - GPG services stuff
Updated all of the UI with the new font and images. Added how to play
option to the main menu. Started work on incorporating sounds. Added
achievement images and various game icons. Login is attempted when the
menu is loaded. Screens added to pop up with sign in/out results. All
current achievements have been catered for.